+++ example/lib/widget/nav_column.dart
@@ -1,23 +1,31 @@
 import 'package:flutter/material.dart';
 import 'package:photo_manager_example/widget/theme_button.dart';
 
+String _defaultBuilder(Widget w) {
+  return w.toStringShort();
+}
+
 class NavColumn extends StatelessWidget {
   const NavColumn({
     super.key,
     required this.children,
+    this.titleBuilder = _defaultBuilder,
   });
 
   final List<Widget> children;
+  final String Function(Widget w) titleBuilder;
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.all(8.0),
-      child: Center(
-        child: Column(
-          children: <Widget>[
-            for (final Widget item in children) buildItem(context, item),
-          ],
+      child: SingleChildScrollView(
+        child: Center(
+          child: Column(
+            children: <Widget>[
+              for (final Widget item in children) buildItem(context, item),
+            ],
+          ),
         ),
       ),
     );
@@ -25,7 +33,7 @@ class NavColumn extends StatelessWidget {
 
   Widget buildItem(BuildContext context, Widget item) {
     return ThemeButton(
-      text: item.toStringShort(),
+      text: titleBuilder(item),
       onPressed: () {
         Navigator.push<void>(
           context,