+++ lib/src/internal/plugin.dart
@@ -84,12 +84,6 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin {
   }
 
   Future<int> getAssetCountFromPath(AssetPathEntity path) async {
-    // Use `assetCount` for Android until we break the API
-    // and migrate to `InternalAssetPathEntity`.
-    if (Platform.isAndroid) {
-      // ignore: deprecated_member_use_from_same_package
-      return path.assetCount;
-    }
     final int result = await _channel.invokeMethod<int>(
       PMConstants.mGetAssetCountFromPath,
       <String, dynamic>{
@@ -255,6 +249,13 @@ class PhotoManagerPlugin with BasePlugin, IosPlugin, AndroidPlugin {
       <String, dynamic>{'ids': ids},
     ) as List<dynamic>;
     return deleted.cast<String>();
+  }
+
+  Future<List<String>> moveToTrash(List<AssetEntity> list) {
+    return _channel.invokeMethod(
+      PMConstants.mMoveToTrash,
+      <String, dynamic>{'ids': list.map((e) => e.id).toList()},
+    ).then((value) => value.cast<String>());
   }
 
   final Map<String, dynamic> onlyAddPermission = <String, dynamic>{